requirements.txt
# CORE DEPENDENCIES - EXACT VERSIONS ONLY
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0

# AI DEPENDENCIES
google-generativeai==0.3.2

# WEB SCRAPING
beautifulsoup4==4.12.2

# DATA ANALYSIS DEPENDENCIES
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# SUPABASE WITH REALTIME FIX - COMPATIBLE VERSIONS
supabase==2.3.4
realtime==1.0.3
websockets==11.0.3
